+++ lib/constants/help.dart
@@ -1,0 +1,19 @@
+class HelpPrints {
+  static final commandsHelp = '''
+------------------------------------------------------------------------------------------------------------
+  -> Commands:
+
+  ->|--------------------|-------------|----------------------------------------|
+    | Command            | Value       | Description                            |
+    |--------------------|-------------|----------------------------------------|
+    | -g (or) --generate | dart        | Generate dart code for keys and values.|
+    |                    | dart-keys   | Generate dart code for keys only.      |
+    |                    | dart-values | Generate dart code for values only.    |
+    |                    | json-values | Generate dart code for values only.    |
+    |--------------------|-------------|----------------------------------------|
+
+  -> For more information vist: https://pub.dev/packages/translations_code_gen
+  -> or visit: https://github.com/youhanasheriff/translations_code_gen
+------------------------------------------------------------------------------------------------------------
+''';
+}