+++ lib/generators/generators.dart
@@ -1,0 +1,3 @@
+export './generate_dart_keys.dart';
+export './generate_dart_values.dart';
+export './generate_json_values.dart';