+++ lib/translations_code_gen.dart
@@ -1,127 +1,256 @@
-/// Generates a Dart code that defines class constants for keys
-/// of a given [Map] of [String] keys to [dynamic] values.
-/// The generated class name will be the same as the key in the
-/// input map, and it will have constants for each key in its
-/// associated value map.
-///
-/// The generated code includes a header indicating that it was
-/// auto-generated, and it includes the necessary ignore directives
-/// to avoid linter warnings for constant identifier names and
-/// camel case types.
-///
-/// The returned [String] contains the generated Dart code.
-///
-/// Example:
-/// ```dart
-/// final Map<String, dynamic> data = {
-///  'App': {
-///   'title': 'Flutter Demo Home Page',
-///  'message': 'You have pushed the button this many times:',
-///   },
-/// };
-///
-/// final dartCode = generateDartCodeKeys(data);
-///
-/// print(dartCode);
-/// // Prints:
-/// // class App {
-/// //   static const String title = "App.title";
-/// //   static const String message = "App.message";
-/// // }
-/// ```
-///
-String generateDartCodeKeys(Map<String, dynamic> data) {
-  final buffer = StringBuffer();
+import 'package:settings_yaml/settings_yaml.dart';
 
-  // Add auto-generated header and ignore directives
-  buffer.writeln('// GENERATED CODE - DO NOT MODIFY BY HAND');
-  buffer.writeln(
-      '// ignore_for_file: constant_identifier_names, camel_case_types');
-  buffer.writeln();
+import 'dart:convert';
+import 'dart:io';
 
-  // Generate the class constants
-  for (var entry in data.entries) {
-    buffer.writeln('class ${entry.key} {');
+import 'constants/constants.dart';
+import 'constants/errors.dart';
+import 'generators/generators.dart';
 
-    for (var field in (entry.value as Map<String, dynamic>).entries) {
-      buffer.write(
-        '  static const String ${field.key} = "${entry.key}.${field.key}";\n',
-      );
-    }
+void translationCodeGen(List<String> arguments) async {
+  String pathToConfiguration = 'translations_code_gen.yaml';
+  String generateMode = SupportedGenerateModes.dart;
 
-    buffer.writeln('}');
-    buffer.writeln();
+  if (!File(pathToConfiguration).existsSync()) {
+    stderr.writeln();
+    stderr.writeln(ConfigErrors.configNotFoundWarning);
+    pathToConfiguration = 'pubspec.yaml';
   }
 
-  return buffer.toString();
+  if (arguments.isNotEmpty) {
+    for (int i = 0; i < arguments.length; i++) {
+      final arg = arguments[i];
+      if (arg.contains(SupportedArguments.supportedArguments[0]) ||
+          arg.contains(SupportedArguments.supportedArguments[1])) {
+        final mode = arg.contains('=') ? arg.split('=').last : arguments[i + 1];
+        if (SupportedGenerateModes.supportedGenerateModes.contains(mode)) {
+          generateMode = mode;
+        } else {
+          stderr.writeln(GenerateErrors.invalidGenerateModeError);
+          exit(1);
+        }
+      }
+    }
+  }
+
+  if (arguments.length > 1) {
+    stderr.writeln(ArgumentsErrors.moreThanOneArgumentError);
+    exit(1);
+  }
+
+  SettingsYaml coonfiguration =
+      SettingsYaml.load(pathToSettings: pathToConfiguration);
+
+  final configExist = coonfiguration.selectorExists('translations_code_gen');
+
+  if (!configExist) {
+    stderr.writeln(ConfigErrors.configNotSpecifiedError);
+    exit(1);
+  }
+
+  String? keysInput =
+      coonfiguration.selectAsMap('translations_code_gen')?['keys']?['input'];
+  String? keysOutput =
+      coonfiguration.selectAsMap('translations_code_gen')?['keys']?['output'];
+  String? valuesInput =
+      coonfiguration.selectAsMap('translations_code_gen')?['values']?['input'];
+  String? valuesOutput =
+      coonfiguration.selectAsMap('translations_code_gen')?['values']?['output'];
+
+  if (keysInput == null ||
+      keysOutput == null ||
+      valuesInput == null ||
+      valuesOutput == null) {
+    stderr.writeln(ConfigErrors.configNotSpecifiedError);
+    exit(1);
+  }
+
+  if (generateMode == SupportedGenerateModes.dart ||
+      generateMode == SupportedGenerateModes.dartKeys) {
+    await _generateDartKeys(
+      keysInput: keysInput,
+      keysOutput: keysOutput,
+    );
+  }
+
+  if (generateMode == SupportedGenerateModes.dart ||
+      generateMode == SupportedGenerateModes.dartValues) {
+    await _generateDartValues(
+      valuesInput: valuesInput,
+      valuesOutput: valuesOutput,
+      generateMode: generateMode,
+    );
+  }
+
+  if (generateMode == SupportedGenerateModes.jsonValues) {
+    await _generateJsonValues(
+      valuesInput: valuesInput,
+      valuesOutput: valuesOutput,
+    );
+  }
 }
 
-/// Generates a Dart code that defines a [Map] of [String] keys to
-/// [String] values for a given [Map] of [String] keys to [dynamic]
-/// values.
-/// The generated [Map] will have the same keys as the input map,
-/// and it will have values that are the concatenation of the key
-/// in the input map and the key in the associated value map.
-/// The generated [Map] will be assigned to a variable named
-/// [lang]Values, where [lang] is the name of the language
-/// specified in the input map.
-/// The generated code includes a header indicating that it was
-/// auto-generated, and it includes the necessary ignore directives
-/// to avoid linter warnings for constant identifier names.
-/// The returned [String] contains the generated Dart code.
-/// Example:
-/// ```dart
-/// final Map<String, dynamic> data = {
-/// 'App': {
-///     'title': 'Flutter Demo Home Page',
-///     'message': 'You have pushed the button this many times:',
-///   },
-/// };
-///
-/// final dartCode = generateDartCodeValues(data, 'en');
-///
-/// print(dartCode);
-/// // Prints:
-/// // const Map<String, String> _app  = {
-/// //   App.title: "Flutter Demo Home Page",
-/// //   App.message: "You have pushed the button this many times:",
-/// // };
-/// //
-/// // final Map<String, String> enValues = {
-/// //   ..._app,
-/// // };
-/// ```
-///
-String generateDartCodeValues(Map<String, dynamic> data, String lang) {
-  final buffer = StringBuffer();
-  final entryKeys = [];
-  buffer.writeln('// GENERATED CODE - DO NOT MODIFY BY HAND');
-  buffer.writeln('// ignore_for_file: constant_identifier_names');
-  buffer.writeln();
-  buffer.writeln("import '../keys.dart';");
-  buffer.writeln();
+Future<void> _generateDartKeys({
+  required String keysInput,
+  required String keysOutput,
+}) async {
+  final inputFileName = keysInput;
+  final outputFileName = keysOutput;
 
-  // Generate the class
-  for (var entry in data.entries) {
-    buffer.write(
-      'const Map<String, String> _${entry.key.toLowerCase()}  = {\n',
-    );
-    entryKeys.add('_${entry.key.toLowerCase()}');
+  // Read the JSON file
+  final file = File(inputFileName);
+  final jsonString = await file.readAsString();
+  final Map<String, dynamic> data = jsonDecode(jsonString);
 
-    for (var field in (entry.value as Map<String, dynamic>).entries) {
-      buffer.writeln(
-        '  ${entry.key}.${field.key}: "${field.value.replaceAll('"', '\\"').replaceAll('\n', '\\n')}",',
-      );
+  // Generate the Dart code
+  final dartCode = generateDartCodeKeys(data);
+
+  // Create the output directory if it does not exist
+  final outputDir = Directory(outputFileName.contains('/')
+      ? outputFileName
+          .split('/')
+          .sublist(0, outputFileName.split('/').length - 1)
+          .join('/')
+      : '.');
+
+  if (!await outputDir.exists()) {
+    await outputDir.create(recursive: true);
+  }
+
+  stderr.writeln();
+  stderr.writeln('GENERATING DART KEYS');
+  stderr.writeln('  -> Generating $outputFileName...');
+
+  // Write the Dart code to a file
+  final outputFile = File(outputFileName);
+  await outputFile.create(recursive: true);
+  await outputFile.writeAsString(dartCode);
+
+  stderr.writeln('DONE!!');
+  stderr.writeln();
+}
+
+Future<void> _generateDartValues({
+  required String valuesInput,
+  required String valuesOutput,
+  required String generateMode,
+}) async {
+  {
+    final inputFileDir = valuesInput[valuesInput.length - 1] == '/'
+        ? valuesInput.substring(0, valuesInput.length - 1)
+        : '$valuesInput/';
+    final outputFileDir = valuesOutput[valuesOutput.length - 1] == '/'
+        ? valuesOutput.substring(0, valuesOutput.length - 1)
+        : '$valuesOutput/';
+
+    // Get a Directory object for the specified path
+    Directory dir = Directory(inputFileDir);
+
+    // Check if the directory exists
+    if (dir.existsSync()) {
+      // List the contents of the directory
+      List<FileSystemEntity> contents = dir.listSync();
+
+      stderr.writeln('GENERATING DART VALUES');
+      // Print the names of the files and directories in the directory
+      for (FileSystemEntity entity in contents) {
+        // Create the output directory if it does not exist
+        // Read the JSON file
+        final file = File(entity.path);
+        final jsonString = await file.readAsString();
+        final Map<String, dynamic> data = jsonDecode(jsonString);
+
+        final lang = entity.path.contains('/')
+            ? entity.path.split('/').last.split('.').first
+            : entity.path.split('.').first;
+
+        // Generate the Dart code
+        final dartCode = generateDartCodeValues(
+          data,
+          lang,
+          generateMode: generateMode,
+        );
+
+        final outputDir = Directory(outputFileDir.contains('/')
+            ? outputFileDir
+                .split('/')
+                .sublist(0, outputFileDir.split('/').length - 1)
+                .join('/')
+            : '.');
+
+        if (!await outputDir.exists()) {
+          await outputDir.create(recursive: true);
+        }
+
+        stderr.writeln('  -> Generating ${outputDir.path}/$lang.dart...');
+
+        // Write the Dart code to a file
+        final outputFile = File('$outputFileDir/$lang.dart');
+        await outputFile.create(recursive: true);
+        await outputFile.writeAsString(dartCode);
+      }
+      stderr.writeln('DONE!!');
+      stderr.writeln();
     }
-    buffer.writeln('};');
-    buffer.writeln();
   }
+}
 
-  buffer.writeln('final Map<String, String> ${lang}Values = {');
-  for (var entryKey in entryKeys) {
-    buffer.writeln('  ...$entryKey,');
+Future<void> _generateJsonValues({
+  required String valuesInput,
+  required String valuesOutput,
+}) async {
+  {
+    final inputFileDir = valuesInput[valuesInput.length - 1] == '/'
+        ? valuesInput.substring(0, valuesInput.length - 1)
+        : '$valuesInput/';
+    final outputFileDir = valuesOutput[valuesOutput.length - 1] == '/'
+        ? valuesOutput.substring(0, valuesOutput.length - 1)
+        : '$valuesOutput/';
+
+    // Get a Directory object for the specified path
+    Directory dir = Directory(inputFileDir);
+
+    // Check if the directory exists
+    if (dir.existsSync()) {
+      // List the contents of the directory
+      List<FileSystemEntity> contents = dir.listSync();
+
+      stderr.writeln('GENERATING JSON VALUES');
+      // Print the names of the files and directories in the directory
+      for (FileSystemEntity entity in contents) {
+        // Create the output directory if it does not exist
+        // Read the JSON file
+        final file = File(entity.path);
+        final jsonString = await file.readAsString();
+        final Map<String, dynamic> data = jsonDecode(jsonString);
+
+        final lang = entity.path.contains('/')
+            ? entity.path.split('/').last.split('.').first
+            : entity.path.split('.').first;
+
+        // Generate the Dart code
+        final jsonStringValue = generateJsonValues(data, lang);
+
+        final outputDir = Directory(outputFileDir.contains('/')
+            ? outputFileDir
+                .split('/')
+                .sublist(0, outputFileDir.split('/').length - 1)
+                .join('/')
+            : '.');
+
+        if (!await outputDir.exists()) {
+          await outputDir.create(recursive: true);
+        }
+
+        stderr.writeln('  -> Generating ${outputDir.path}/$lang.json...');
+
+        // Write the Dart code to a file
+        final outputFile = File('$outputFileDir/$lang.json');
+        await outputFile.create(recursive: true);
+        await outputFile.writeAsString(jsonStringValue);
+      }
+      stderr.writeln('DONE!!');
+      stderr.writeln();
+    }
   }
-  buffer.writeln('};');
-
-  return buffer.toString();
 }